project_management_system/requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.11.17
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.7
django-environ==0.11.2
django-filter==23.3
django-import-export==3.3.1
djangorestframework==3.14.0
environ==1.0
et-xmlfile==1.1.0
fiona==1.9.5
geopandas==0.14.1
import-export==0.3.1
MarkupPy==1.14
numpy==1.26.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
psycopg2==2.9.9
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.8.0
tzdata==2023.3
xlrd==2.0.1
xlwt==1.3.0
celery==5.2.7
redis==4.3.3
mqp==5.1.1
asgiref==3.7.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.4
certifi==2023.7.22
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
Django==4.2.4
django-debug-toolbar==4.2.0
django-filter==23.2
django-leaflet==0.29.0
djangorestframework==3.14.0
Faker==19.6.1
Fiona==1.9.4.post1
geopandas==0.13.2
kombu==5.3.2
Markdown==3.4.4
numpy==1.25.2
packaging==23.1
pandas==2.1.0
Pillow==10.0.0
prompt-toolkit==3.0.39
psycopg2==2.9.7
psycopg2-binary==2.9.7
pyproj==3.6.0
pyshp==2.3.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.6
XlsxWriter==3.1.2
